refactor(front_canvas): extract canvas-helpers and groups modules; wire into layers panel
